renamed file to useDaoDetails
