+++ packages/web-app/src/hooks/useDaoDetails.tsx
@@ -1,10 +1,11 @@
 import {Client, DaoDetails} from '@aragon/sdk-client';
+import {isAddress} from '@ethersproject/address';
 import {useQuery} from '@tanstack/react-query';
-import {useCallback, useEffect} from 'react';
+import {useCallback, useEffect, useMemo} from 'react';
 import {useNavigate, useParams} from 'react-router-dom';
 
 import {useNetwork} from 'context/network';
-import {resolveDaoAvatarIpfsCid} from 'utils/library';
+import {resolveDaoAvatarIpfsCid, toDisplayEns} from 'utils/library';
 import {NotFound} from 'utils/paths';
 import {useClient} from './useClient';
 
@@ -41,7 +42,10 @@ export const useDaoQuery = (
   const {client, network: clientNetwork} = useClient();
 
   // if network is unsupported this will be caught when compared to client
-  const queryNetwork = networkUrlSegment ?? network;
+  const queryNetwork = useMemo(
+    () => networkUrlSegment ?? network,
+    [network, networkUrlSegment]
+  );
 
   // make sure that the network and the url match up with client network before making the request
   const enabled =
@@ -57,7 +61,7 @@ export const useDaoQuery = (
     select: addAvatarToDao,
     enabled,
     refetchOnWindowFocus: false,
-    refetchInterval: () => (refetchInterval ? refetchInterval : false),
+    refetchInterval,
   });
 };
 
@@ -74,16 +78,29 @@ export const useDaoDetailsQuery = () => {
   const apiResponse = useDaoQuery(daoAddressOrEns);
 
   useEffect(() => {
-    // navigate to 404 if the DAO is not found or there is some sort of error
-    if (
-      // no live dao
-      apiResponse.isFetched &&
-      (apiResponse.error || apiResponse.data === null)
-    ) {
-      navigate(NotFound, {
-        replace: true,
-        state: {incorrectDao: daoAddressOrEns},
-      });
+    if (apiResponse.isFetched) {
+      // navigate to 404 if the DAO is not found or there is some sort of error
+      if (apiResponse.error || apiResponse.data === null) {
+        navigate(NotFound, {
+          replace: true,
+          state: {incorrectDao: daoAddressOrEns},
+        });
+      }
+
+      //navigate to url with ens domain
+      else if (
+        isAddress(daoAddressOrEns as string) &&
+        toDisplayEns(apiResponse.data?.ensDomain)
+      ) {
+        const segments = location.pathname.split('/');
+        const daoIndex = segments.findIndex(
+          segment => segment === daoAddressOrEns
+        );
+        if (daoIndex !== -1 && apiResponse.data?.ensDomain) {
+          segments[daoIndex] = apiResponse.data.ensDomain;
+          navigate(segments.join('/'));
+        }
+      }
     }
   }, [
     apiResponse.data,
